댓글 삭제하기

+++ src/main/java/shop/mtcoding/blog/reply/ReplyJPARepository.java
@@ -1,7 +1,8 @@
 package shop.mtcoding.blog.reply;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyJPARepository extends JpaRepository<Reply, Long> {
+public interface ReplyJPARepository extends JpaRepository<Reply, Integer> {
 
 }
